pyproject.toml
[project]
name = "arango-cypher-py"
version = "0.0.0"
description = "Python-native Cypher to AQL transpiler for ArangoDB."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
  "python-arango>=8.1.1",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.0",
  "antlr4-python3-runtime",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "antlr4-tools",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: requires a running ArangoDB (typically via docker compose)",
  "tck: openCypher TCK harness and runners",
]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["arango_query_core", "arango_cypher"]

